Rename mapping_tools.hpp
