+++ src/advection/spline_polar_foot_finder.hpp
@@ -4,11 +4,11 @@
 
 #include "circular_to_cartesian.hpp"
 #include "combined_mapping.hpp"
+#include "coord_transformation_tools.hpp"
 #include "ddc_alias_inline_functions.hpp"
 #include "ddc_aliases.hpp"
 #include "geometry_pseudo_cartesian.hpp"
 #include "ipolar_foot_finder.hpp"
-#include "mapping_tools.hpp"
 #include "vector_index_tools.hpp"
 #include "vector_mapper.hpp"
