+++ src/coord_transformations/README.md
@@ -195,4 +195,4 @@ The tools are:
 - InvJacobianOPoint : this tool calculates the inverse Jacobian matrix at the O-point on the specified coordinate system.
 - MetricTensorEvaluator : this tool calculates the metric tensor associated with a coordinate transformation.
 - VectorMapper : this tool helps when converting vectors stored in a `VectorField` from one coordinate system to another.
-- other static analysis tools found in `mapping_tools.hpp`
+- other static analysis tools found in `coord_transformation_tools.hpp`
