+++ src/coord_transformations/cartesian_to_circular.hpp
@@ -3,8 +3,8 @@
 
 #include <ddc/ddc.hpp>
 
+#include "coord_transformation_tools.hpp"
 #include "ddc_aliases.hpp"
-#include "mapping_tools.hpp"
 #include "tensor.hpp"
 #include "view.hpp"
