+++ src/coord_transformations/discrete_to_cartesian.hpp
@@ -4,9 +4,9 @@
 
 #include <ddc/ddc.hpp>
 
+#include "coord_transformation_tools.hpp"
 #include "ddc_alias_inline_functions.hpp"
 #include "ddc_aliases.hpp"
-#include "mapping_tools.hpp"
 #include "math_tools.hpp"
 #include "tensor.hpp"
 #include "view.hpp"
