+++ src/coord_transformations/inverse_jacobian_matrix.hpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
+#include "coord_transformation_tools.hpp"
 #include "ddc_aliases.hpp"
-#include "mapping_tools.hpp"
 #include "math_tools.hpp"
 #include "view.hpp"
