+++ src/interpolation/polar_splines/polar_bsplines.hpp
@@ -7,9 +7,9 @@
 
 #include "bernstein.hpp"
 #include "cartesian_to_barycentric.hpp"
+#include "coord_transformation_tools.hpp"
 #include "ddc_helper.hpp"
 #include "discrete_to_cartesian.hpp"
-#include "mapping_tools.hpp"
 #include "polar_spline.hpp"
 #include "view.hpp"
