+++ src/multipatch/connectivity/onion_patch_locator.hpp
@@ -5,8 +5,8 @@
 
 #include <ddc/ddc.hpp>
 
+#include "coord_transformation_tools.hpp"
 #include "ddc_aliases.hpp"
-#include "mapping_tools.hpp"
 #include "multipatch_type.hpp"
 #include "types.hpp"
