+++ src/quadrature/volume_quadrature_nd.hpp
@@ -8,9 +8,9 @@
 #pragma once
 #include <ddc/ddc.hpp>
 
+#include "coord_transformation_tools.hpp"
 #include "ddc_alias_inline_functions.hpp"
 #include "ddc_aliases.hpp"
-#include "mapping_tools.hpp"
 #include "quadrature.hpp"
